refactor node_manager api design

+++ src/node_manager/services/service_environments.py
@@ -1,7 +1,10 @@
+from uuid import UUID
+
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from node_manager.schemas import schemas_environments
-from node_manager.crud import crud_environments, crud_platforms
+from node_manager.crud import crud_environments
+from node_manager.utils import is_uuid
 from schemas.response_schemas import GenericResponseModel
 
 
@@ -9,8 +12,24 @@ async def get_all_environments(
         platform_name: str,
         session: AsyncSession
 ) -> GenericResponseModel:
-    data = await crud_platforms.get_platform_by_name(platform_name=platform_name, session=session)
-    return GenericResponseModel(data=data, total=len(data.environments))
+    data = await crud_environments.get_platform_environments(
+        platform_name=platform_name,
+        session=session
+    )
+    return GenericResponseModel(data=data, total=len(data))
+
+
+async def get_environment(
+        platform_name: str,
+        environment: str | UUID,
+        session: AsyncSession
+) -> GenericResponseModel:
+    data = await crud_environments.get_environment(
+        environment=environment,
+        platform_name=platform_name,
+        session=session
+    )
+    return GenericResponseModel(data=data, total=1)
 
 
 async def create_environment(
@@ -28,12 +47,20 @@ async def create_environment(
 
 async def delete_environment(
         platform_name: str,
-        environment_id: int,
+        environment: str | UUID,
         session: AsyncSession
 ) -> GenericResponseModel:
-    data = await crud_environments.delete_environment_by_id(
+    if is_uuid(environment):
+        data = await crud_environments.delete_environment_by_id(
+            platform_name=platform_name,
+            environment_id=environment,
+            session=session
+        )
+        return GenericResponseModel(data=data)
+
+    data = await crud_environments.delete_environment_by_name(
         platform_name=platform_name,
-        environment_id=environment_id,
+        environment_name=environment,
         session=session
     )
     return GenericResponseModel(data=data)
